refactore: folder structure
