+++ src/pages/admin/RealtimeSeats.tsx
@@ -6,8 +6,8 @@ import TopNav from "@components/layout/headers/TopNav";
 import PlaySessionPicker from "@components/layout/headers/PlaySessionPicker";
 import AdminSeatInfo from "@components/common/informations/AdminSeatInfo";
 import AudienceInfo from "@components/pages/admin/realtimeSeats/AudienceInfo";
-import AdminSeatMap from "@components/common/buttons/SeatMap/AdminSeatMap_Riveract";
-// import AdminSeatMap from '../../components/button/SeatMap/AdminSeatMap_Kwangwoon';
+import AdminSeatMap from "@components/pages/admin/seatsComponent/AdminSeatMap_Riveract";
+// import AdminSeatMap from '@components/pages/admin/seatsComponent/AdminSeatMap_Kwangwoon';
 
 import backIcon from "@assets/images/left_arrow.png";
 import refreshIcon from "@assets/images/refresh2_icon.png";
